improved database safety + minor bug fixes Add varsions to all tables, added new static class for concurency checking, replaced all '_context.SaveChanges()' with new 'SafeSave' method on static class

+++ wms-praktyki-yosi-api/wms-praktyki-yosi-api/Controllers/DocumentController.cs
@@ -11,18 +11,23 @@ namespace wms_praktyki_yosi_api.Controllers
     [Authorize]
     [ApiController]
     [ServiceFilter(typeof(ValidationFilterAttribute))]
-    public class DocumetController : ControllerBase
+    public class DocumentController : ControllerBase
     {
         private readonly IDocumentService _service;
+        private readonly ICustomAuthorizationService _authorizationService;
 
-        public DocumetController(IDocumentService service)
+        public DocumentController(IDocumentService service, ICustomAuthorizationService authorizationService)
         {
             _service = service;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
         public async Task<ActionResult<List<DocumentDto>>> GetAllDocuments([FromQuery]GetRequestQuery query)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             var documents = _service.GetAllDocuments(query);
             return Ok(documents);
         }
@@ -30,27 +35,41 @@ namespace wms_praktyki_yosi_api.Controllers
         [HttpPost]
         public async Task<ActionResult> AddDocument([FromBody] AddDocumentDto dto)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             var createdId = _service.AddDocument(dto);
-            return Created("api/document/" + createdId, null);
+            return Created("/api/documents/" + createdId, null);
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<DetailedDocumentDto>> GetDocumentDetails([FromRoute]string id)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             var document = _service.GetDocumentDetails(id);
             return Ok(document);
         }
 
         [HttpDelete("{id}")]
+        [Authorize(Roles = "Admin")]
         public async Task<ActionResult> DeleteDocument([FromRoute] string id)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             _service.DeleteDocument(id);
             return Ok();
         }
 
         [HttpPost("{id}/markasfinished")]
+        [Authorize(Roles = "Admin,Moderator")]
         public async Task<ActionResult> MarkDocumentAsFinished([FromRoute] string id, [FromBody] bool finished)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             _service.MarkDocumentAsFinished(id, finished);
             return Ok();
         }
@@ -60,6 +79,9 @@ namespace wms_praktyki_yosi_api.Controllers
             [FromRoute] string id,
             [FromBody] DocumentVisitLocationDto location)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             _service.VisitLocation(id, location);
             return Ok();
         }
@@ -67,30 +89,45 @@ namespace wms_praktyki_yosi_api.Controllers
         [HttpGet("{id}/items")]
         public async Task<ActionResult> GetDocumentItems([FromRoute] string id, [FromQuery]GetRequestQuery query)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             var items = _service.GetDocumentItems(id, query);
             return Ok(items);
         }
 
         [HttpPost("{id}/items")]
+        [Authorize(Roles = "Admin,Moderator")]
         public async Task<ActionResult> AddItemToDocument([FromRoute] string id, [FromBody] AddDocumentItemDto item)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             _service.AddItemToDocument(id, item);
             return Ok();
         }
 
-        [HttpDelete("{id}/items/{productId}")]
-        public async Task<ActionResult> DeleteItemFromDocument([FromRoute] string id, [FromRoute] int productId)
+        [HttpDelete("{id}/items/{itemId}")]
+        [Authorize(Roles = "Admin,Moderator")]
+        public async Task<ActionResult> DeleteItemFromDocument([FromRoute] string id, [FromRoute] string itemId)
         {
-            _service.DeleteDocumentItem(id, productId);
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
+            _service.DeleteDocumentItem(id, itemId);
             return Ok();
         }
 
         [HttpPut("{id}/items/{productId}")]
+        [Authorize(Roles = "Admin,Moderator")]
         public async Task<ActionResult> UpdateItemInDocument(
             [FromRoute] string id,
             [FromRoute] int productId,
             [FromBody] EditDocumentItemDto item)
         {
+            if (!await _authorizationService.UserIsAuthorized(User))
+                return Unauthorized();
+
             _service.UpdateItemInDocument(id, productId, item);
             return Ok();
         }
